(3) Simulation Definition

FLOWUnsteady.Simulation — Type

Simulation{V<:AbstractVehicle, M<:AbstractManeuver, R<:Real}(vehicle::V,
                        maneuver::M, Vref::R, RPMref::R, ttot::R, optargs...)

Simulation interface. This type carries the simulation's options and connects vehicle and maneuver together.

ARGUMENTS

vehicle : Vehicle

maneuver : Maneuver

Vref : Reference velocity for the maneuver

RPMref : Reference RPM for the maneuver

ttot : Total time in which to perform the maneuver

OPTIONAL ARGUMENTS

Vinit = zeros(3) : Initial vehicle velocity

Winit = zeros(3) : Initial vehicle angular velocity

State variables

t::Real : Time of current step

nt::Int : Current time step number

FLOWUnsteady.save_vtk — Function

save_vtk(self::AbstractVehicle, prefix; path="", optargs...)

Output VTK files with vehicle geometry and solution fields.

save_vtk(sim::Simulation, prefix; path="", save_wopwopin=false, optargs...)

Output VTK files with vehicle geometry and solution fields. The file names will have the prefix prefix, and will be saved in the directory path. If save_wopwopin=true, it will also generate PSU-WOPWOP input files that can be used to run the acoustic analysis (see run_noise_wopwop).