Vehicle types

FLOWUnsteady.UVLMVehicle — Type

UVLMVehicle{N, M, R}(system; tilting_systems, rotors_systems,
                                        vlm_system, wake_system, grids)

Type handling all geometries and subsystems that define a vehicle made out of FLOWVLM components (Wing, WingSystem, Rotor).

ARGUMENTS

system::FLOWVLM.WingSystem: System of all FLOWVLM objects. This system is considered as the entire vehicle. Not all components in this system will be solved, but they will all be rotated and translated according to the maneuver.

OPTIONAL ARGUMENTS

tilting_systems::NTuple{N, FLOWVLM.WingSystem}: Tuple of all FLOWVLM tilting objects, where tilting_systems[i] contains the i-th FLOWVLM system of lifting surfaces and rotors that tilt together.

rotors_systems::NTuple{M, Array{vlm.Rotor}}: Tuple of groups of Rotors that share a common RPM.

vlm_system::FLOWVLM.WingSystem: System of all FLOWVLM objects to be solved through the VLM solver.

wake_system::FLOWVLM.WingSystem: System of all FLOWVLM objects that will shed a VPM wake.

grids::Array{gt.GridTypes}: Array of grids that will be translated and rotated along with system.

State variables

V::Vector : Current vehicle velocity

W::Vector : Current vehicle angular velocity

prev_data::Array{Any} : Information about previous step

grid_O::Vector{Vector} : Origin of every grid

FLOWUnsteady.VLMVehicle — Type

Alias for FLOWUnsteady.UVLMVehicle

FLOWUnsteady.QVLMVehicle — Type

QVLMVehicle{N, M, R}(system; optargs...)

Same than FLOWUnsteady.UVLMVehicle but replacing the VPM wake with a semi-infinite rigid VLM wake, making the simulation quasi-ssteady.

NOTE: For the solver to work correctly, all components in wake_system (if any) need to be also components of vlm_system.

NOTE 2: It is recommended that wake_system doesn't include any Rotor object. Otherwise, blades will generate a wake going straight out of every blade trailing edge pointing oposite to the direction of rotation instead of generating a streamtube.