Relaxation Scheme

FLOWVPM.Relaxation — Type

`Relaxation(relax, nsteps_relax, rlxf)`

Defines a relaxation method implemented in the function relax(rlxf::Real, p) where p is particle, rlxf is the relaxation factor between 0 and 1, with 0 == no relaxation, and 1 == full relaxation. The simulation is relaxed every nsteps_relax steps.

FLOWVPM.relax_pedrizzetti — Function

`relax_Pedrizzetti(rlxf::Real, p)`

Relaxation scheme where the vortex strength is aligned with the local vorticity.

FLOWVPM.relax_correctedpedrizzetti — Function

`relax_correctedPedrizzetti(rlxf::Real, p)`

Relaxation scheme where the vortex strength is aligned with the local vorticity. This version fixes the error in Pedrizzetti's relaxation that made the strength to continually decrease over time. See notebook 20200921 for derivation.