Functions Contained in FLOWUnsteady

FLOWUnsteady.VLM2VPM — Method

Receives the FLOWVLM object system (Wing/WingSystem/Rotor), and adds vortex particles to the particle field pfield at each trailing edge position where an infinite vortex starts. dt indicates the length of the interval of time that the vortex shedding represents.

Give it a previous system to detect differences in circulation and add unsteady particles.

FLOWUnsteady.Vvpm_on_Xs — Method

Returns the velocity induced by particle field on every position Xs

FLOWUnsteady._Vkinematic_rotor — Method

Returns the local translational velocity of every midpoint in the ri-th rotor in the si-th system formatted as the output of _get_midXs().

FLOWUnsteady._Vkinematic_vlm — Method

Returns the local translational velocity of every control point in vlm_system.

FLOWUnsteady._Vkinematic_wake — Method

Returns the local translational velocity of every control point in wake_system.

FLOWUnsteady._check — Method

Checks that vehicle and maneuver are compatible.

FLOWUnsteady._extraVinf1 — Method

Function for FLOWVLM to fetch and add extra velocities to what every control point imposes as the boundary condition

FLOWUnsteady._format_blades — Method

Receives the output of `parse_midXs(rotor_systems, midXs, si, ri)` and

outputs the array formatted as blades.

FLOWUnsteady._get_midX — Method

Returns the i-th bound-vortex midpoint of the j-th blade in rotor

FLOWUnsteady._get_midXs — Method

Returns all bound-vortex midpoints in an array of rotors

FLOWUnsteady._get_midXs — Method

Returns all bound-vortex midpoints of the j-th blade in rotor

FLOWUnsteady._get_midXs — Method

Returns all bound-vortex midpoints in rotor

FLOWUnsteady._get_midXs — Method

Returns all bound-vortex midpoints in multiple rotor systems

FLOWUnsteady._parse_midXs — Method

Receives an array midXs formatted as the output of _get_midXs(rotor_systems) and returns the section of the array that corresponds to the ri-th rotor in the si-th system (formatted linearly).

FLOWUnsteady.add_dV — Method

`add_dV(self::AbstractVehicle, dV)`

Increments the velocity of the vehicle by dV.

FLOWUnsteady.add_dW — Method

`add_dW(self::AbstractVehicle, dV)`

Increments the angular velocity of the vehicle by dW.

FLOWUnsteady.calc_ReD — Method

Reynolds number at radial station r (m)

FLOWUnsteady.calc_aerodynamicforce — Method

Calculates the aerodynamic force at each element of a VLM system using it's

current Gamma solution, the velocity induced by the particle field pfield, and the kinematic velocity calculated from prev_vlm_system. It saves the force as the field vlm_system.sol["Ftot"]

FLOWUnsteady.calc_cbar — Method

Returns the mean chord of the blade of the given rotor

FLOWUnsteady.calc_dV — Method

`calc_dV(maneuver::AbstractManeuver, vehicle::Vehicle, t, dt, ttot, Vref)`

Returns the change in velocity dV=[dVx, dVy, dVz] (m/s) of vehicle performing maneuver at time t (s) after a time step dt (s). Vref and tot are the reference velocity and the total time at which this maneuver is being performed, respectively. dV is in the global reference system.

FLOWUnsteady.calc_dW — Method

`calc_dw(maneuver::AbstractManeuver, vehicle::Vehicle, t, dt, Vref, ttot)`

Returns the change in angular velocity dW=[dWx, dWy, dWz] (about global axes, in radians) of vehicle performing maneuver at time t (s) after a time step dt (s). ttot is the total time at which this maneuver is to be performed.

FLOWUnsteady.calc_n — Method

Given a diameter-based Reynolds number $\mathrm{Re}_D(r)$ at a certain

radial position r and an advance ratio J, return the rotational speed

FLOWUnsteady.decompose — Method

Receives an array of vectors Fs and decomposes them into components in the ihat, jhat, khat unitary orthogonal directions (khat is deduced).

FLOWUnsteady.generate_monitor_rotors — Method

`generate_monitor_rotors( rotors::Array{vlm.Rotor, 1}, J_ref::Real,

rhoref::Real, RPMref::Real, nsteps_sim::Int)`

Generate monitor of load distribution and performance of a rotor system. J_ref and rho_ref are the reference advance ratio and air density used for calculating propulsive efficiency and coefficients. RPM_ref is the reference RPM used to estimate the age of the wake. nsteps_sim is the number of time steps by the end of the simulation (used for generating the color gradient).

FLOWUnsteady.generate_monitor_wing — Method

`generate_monitor_wing(wing, Vinf::Function, b_ref::Real, ar_ref::Real,

rhoref::Real, qinfref::Real, nsteps_sim::Int)`

Generate monitor of load distribution and performance of a wing.

FLOWUnsteady.generate_rotor — Method

Generates the Rotor object. pitch is the pitch of the blades in degrees, n is the number of lattices in the VLM.

ReD is the diameter Reynolds number based on rotational speed calculated as ReD = (omegaR)(2*R)/nu, and Matip is the rotational+freestream Mach number at the tip. This values are used for calculate airfoil properties through XFOIL (chord Reynolds), so ignore them if airfoil properties are prescribed.

Give it altReD = [RPM, J, nu], and it'll calculate the chord-based Reynolds accounting for the effective velocity at every station ignoring ReD (this is more accurate, but not needed).

NOTE: If Matip is different than zero while running XFOIL, remember to deactive compressibility corrections when running vlm.solvefromCCBlade() by giving it sound_spd=nothing

FLOWUnsteady.generate_static_particle_fun — Method

Returns a function that generates an array of particles representing the surface of the vehicle as a collection of vortex particles.

FLOWUnsteady.generate_vtkliftinglines — Method

Generate lifting lines for acoustic compact patches of rotors

FLOWUnsteady.generate_wopwoploading — Method

Generates loading files for PSU-WOPWOP from the indicated simulation in

read_path using the steps nums. If woptype=Constant, only one step must be specified; if woptype=Periodic, it is assumed that the specified steps make one and only one full revolution; if woptype=Aperiodic, it is assumed that the specified steps make one and only one full revolution. This function outputs the load per unit length in the global coordinate system if periodic, or in the blade coordinate system if constant.

NOTE: If periodic, there should be steps from 0 degrees up to 360 degrees (or up to whether the periodic angle is) where nums[1] matches the same conditions than nums[end] (rotor simulations with this module will do this by default). The period of one revolution should be indicated under period.

NOTE2: dt must be the same time step than specified in the PSU-WOPWOP driver.

FLOWUnsteady.get_RPM — Method

`get_RPM(maneuver::AbstractManeuver, i::Int, t::Real)`

Returns the normalized RPM of the i-th rotor system at the non-dimensional time t.

FLOWUnsteady.get_RPMs — Method

`get_RPMs(maneuver::AbstractManeuver, t::Real)`

Returns the normalized RPM of every rotor systems at the non-dimensional time t.

FLOWUnsteady.get_angle — Method

`get_angle(maneuver::AbstractManeuver, i::Int, t::Real)`

Returns the angle (in degrees) of the i-th tilting system at the non-dimensional time t.

FLOWUnsteady.get_angles — Method

`get_angles(maneuver::AbstractManeuver, t::Real)`

Returns the angle (in degrees) of every tilting systems at the non-dimensional time t.

FLOWUnsteady.get_nrtrsys — Method

`get_nrtrsys(self::AbstractManeuver)`

Return number of rotor systems.

FLOWUnsteady.get_nrtrsys — Method

`get_nrtrsys(self::AbstractVehicle)`

Return number of rotor systems.

FLOWUnsteady.get_ntltsys — Method

`get_ntltsys(self::AbstractManeuver)`

Return number of tilting systems.

FLOWUnsteady.get_ntltsys — Method

`get_ntltsys(self::AbstractVehicle)`

Return number of tilting systems.

FLOWUnsteady.nextstep_kinematic — Method

`nextstep_kinematic(self::AbstractVehicle, dt::Real)`

Translates and rotates the vehicle in a time step dt according to current linear and angular velocity.

FLOWUnsteady.nextstep_kinematic — Method

`nextstep_kinematic(self::Simulation, dt::Real)`

Takes a kinematic time step dt where the new velocity and angular velocity is calculated and the vehicle is translated and rotated according to it. It also updates the tilt angle and RPM of every system.

NOTE: No solver is run in this process, rather it uses the current aerodynamic solution to calculate acceleration and moment.

FLOWUnsteady.plot_maneuver — Method

`plot_maneuver(maneuver::KinematicManeuver; ti::Real=0, tf::Real=1,

vis_nsteps=300, figname="maneuver", tstages=[])`

Plots the kinematics and controls of a KinematicManeuver.

FLOWUnsteady.precalculations — Method

Precalculations before calling the solver.

Calculates kinematic velocity on VLM an adds them as a solution field

FLOWUnsteady.precalculations — Method

`precalculations(self, pfield, t, dt)`

Precalculations before calling the solver and before shedding trialing wake.

FLOWUnsteady.precalculations — Method

Precalculations before calling the solver.

FLOWUnsteady.rotate_observers — Method

Rotate observer(s) as the BPM code assumes wind is in the x-y plane. (So for a rotor in hover, the rotors are offset by 90 deg. We can instead rotate the observers to get the same results.)

θ is in degrees.

FLOWUnsteady.run_noise_bpm — Method

Calculates broadband noise from the Brooks, Pope and Marcolini method using the BPM.jl package.

The function writes the data to a file that can be read in postprocessing with FLOWNoise.readwopwopoutput(). It writes three files: SPL (non-weighted), A-weighted SPL, and OASPL.

FLOWUnsteady.run_noise_wopwop — Method

Given the path of a FLOWUnsteady simulation, it runs the noise analysis on the rotors of the simulation. It uses PSU-WOPWOP to calculate the tonal noise from thickness and loading sources on the geometry and aerodynamic loading generated by FLOWUnsteady.

NOTE: This function will call the PSU-WOPWOP binary indicated through wopwopbin. This binary is not included with FLOWUnsteady and must be provided by the user. This function has been tested on PSU-WOPWOP v3.4.2.

NOTE2: Make sure that the simulation was run with nsteps_save=1, otherwise the time in PSU-WOPWOP will get messed up.

FLOWUnsteady.save_vtk — Method

`save_vtk(self::AbstractVehicle, filename; path=nothing, num=nothing,

optargs...)`

Output VTK files with vehicle geometry and solution fields.

FLOWUnsteady.set_V — Method

`set_V(self::AbstractVehicle, V)`

Set current vehicle velocity to V.

FLOWUnsteady.set_W — Method

`set_W(self::AbstractVehicle, W)`

Set current vehicle angular velocity to W.

FLOWUnsteady.shed_wake — Method

Shed VPM wake. If unsteady_shedcrit<=0 it sheds wake due to loading distribution, otherwise it will shed wake due to unsteady loading.

FLOWUnsteady.tilt_systems — Method

`nextstep_kinematic(self::AbstractVehicle, angles)`

Tilts every tilting system of this vehicle into its corresponding new angle, where angles[i]=[Ax, Ay, Az] is the new angle of the i-th tilting system (in degrees).

FLOWUnsteady.visualize_kinematics — Method

`visualize_kinematics(sim::Simulation, nsteps::Int, save_path::String)`

Generate VTKs of the kinematics of this simulation.

FLOWUnsteady.AbstractManeuver — Type

`AbstractManeuver{N, M}`

N indicates the number of tilting systems in this maneuver, while and M indicates the number of rotor systems.

Every implementation of AbstractManeuver must have the properties:

angle::NTuple{N, Function} where angle[i](t) returns the angle of the i-th tilting system at time t (t is nondimensionalized by the total time of the maneuver, from 0 to 1, beginning to end).

RPM::NTuple{M, Function} where RPM[i](t) returns the normalized RPM of the i-th rotor system at time t. This RPM values are normalized by the an arbitrary RPM value (usually RPM in hover or cruise).

FLOWUnsteady.AbstractVLMVehicle — Type

`AbstractVLMVehicle{N, M, R} <: AbstractVehicle{N, M, R}`

Implementations of AbstractUVLMVehicle are expected to have the following fields:

Required inputs

system::vlm.WingSystem

Optional inputs

tilting_systems::NTuple{N, vlm.WingSystem}

rotor_systems::NTuple{M, Array{vlm.Rotor, 1}}

vlm_system::vlm.WingSystem

wake_system::vlm.WingSystem

grids::Array{gt.GridTypes, 1}

Internal properties

V::Array{R, 1} # Current vehicle velocity

W::Array{R, 1} # Current vehicle angular velocity

prev_data::Array{Any, 1} # Information about previous step

grid_O::Array{Array{R, 1}, 1} # Origin of every grid

They also need to implement the following functions required by AbstractVehicle:

shed_wake(...)

generate_static_particle_fun(...)

save_vtk(...)

See the documentation and code of AbstractVehicle for more details.

In general, in order for implementation to work correctly, it is required that the components of wake_system are also components of either vlm_system or rotor_systems, and that none of the components of rotor_systems are also components of vlm_system.

FLOWUnsteady.AbstractVehicle — Type

`AbstractVehicle{N, M, R}`

Type handling all geometries and subsystems that define a flight vehicle.

N indicates the number of tilting systems in this vehicle, while and M indicates the number of rotor systems. R is a Real type.

Implementations must have the following properties: * V::Array{R, 1} : Current velocity vector of the vehicle in the global coordinate system. * W::Array{R, 1} : Current angular velocity vector of the vehicle in the global coordinate system.

FLOWUnsteady.KinematicManeuver — Type

`KinematicManeuver(angle, RPM, Vvehicle, anglevehicle)`

A vehicle maneuver where the kinematic are prescribed.

ARGUMENTS

angle::NTuple{N, Function} where angle[i](t) returns the angles [Ax, Ay, Az] (in degrees)of the i-th tilting system at time t (t is nondimensionalized by the total time of the maneuver, from 0 to 1, beginning to end).

RPM::NTuple{M, Function} where RPM[i](t) returns the normalized RPM of the i-th rotor system at time t. This RPM values are normalized by the an arbitrary RPM value (usually RPM in hover or cruise).

Vvehicle::Function where Vvehicle(t) returns the normalized vehicle velocity [Vx, Vy, Vz] at the normalized time t. Velocity is normalized by a reference velocity (typically, cruise velocity).

anglevehicle::Function where anglevehicle(t) returns the angles [Ax, Ay, Az] (in degrees) of the vehicle relative to the global coordinate system at the normalized time t.

FLOWUnsteady.QVLMVehicle — Type

`QVLMVehicle(system; optargs...)`

Type handling all geometries and subsystems that define a flight vehicle made out of VLM (Wing, WingSystem, Rotor) components.

NOTE: For the solver to work correctly, all components in wake_system (if any) need to be also components of vlm_system.

NOTE: It is recommended that wake_system doesn't include any Rotor object. Otherwise, blades will generate a wake going straight out of every blade trailing edge pointing oposite to the direction of rotation instead of generating a streamtube.

ARGUMENTS

system::vlm.WingSystem: System of all FLOWVLM objects. This system is considered as the entire vehicle. Not all components in this system will be solved, but they will all be rotated and translated during maneuver.

OPTIONAL ARGUMENTS

tilting_systems::Tuple(vlm.WingSystem, ...): Tuple of all FLOWVLM tilting objects, where tilting_systems[i] contains the i-th FLOWVLM system of lifting surfaces and rotors that tilt together.

rotors_systems::Tuple(Array{vlm.Rotor,1}, ...): Tuple of groups of Rotors that share a common RPM.

vlm_system::vlm.WingSystem: System of all FLOWVLM objects to be solved through the VLM solver.

wake_system::vlm.WingSystem: System of all FLOWVLM objects that will have a rigid, semi-infinite VLM wake.

grids::Array{gt.GridTypes, 1}: Array of grids that will be translated and rotated along with system.

FLOWUnsteady.Simulation — Type

`Simulation(vehicle, maneuver, Vref::Real, RPMref::R, ttot:R)`

Simulation interface. This type carries the simulation's options and connects vehicle and maneuver together.

FLOWUnsteady.UVLMVehicle — Type

`UVLMVehicle(system; optargs...)`

Type handling all geometries and subsystems that define a flight vehicle made out of VLM (Wing, WingSystem, Rotor) components.

ARGUMENTS

system::vlm.WingSystem: System of all FLOWVLM objects. This system is considered as the entire vehicle. Not all components in this system will be solved, but they will all be rotated and translated during maneuver.

OPTIONAL ARGUMENTS

tilting_systems::Tuple(vlm.WingSystem, ...): Tuple of all FLOWVLM tilting objects, where tilting_systems[i] contains the i-th FLOWVLM system of lifting surfaces and rotors that tilt together.

rotors_systems::Tuple(Array{vlm.Rotor,1}, ...): Tuple of groups of Rotors that share a common RPM.

vlm_system::vlm.WingSystem: System of all FLOWVLM objects to be solved through the VLM solver.

wake_system::vlm.WingSystem: System of all FLOWVLM objects that will shed a VPM wake.

grids::Array{gt.GridTypes, 1}: Array of grids that will be translated and rotated along with system.