Functions Contained in FLOWVLM

FLOWVLM.FLOWVLM2OCCBlade — Method

Returns a CCBlade's rotor object corresponding to the i-th blade. The r position used to generate the object are the control points, Rhub is the minimum radius given when initializing the FLOWVLM Rotor object, and Rtip is the maximum. The precone given to CCBlade is 0 since precone it is expected to be already accounted for in the geometry given by the user when initializing the FLOWVLM Rotor.

It applies 3D corrections to the airfoil polars given for initializing the FLOWVLM Rotor, hence the RPMs for those corrections must be specified.

WARNING: This function will use the Inflow field, hence make sure they are updated.

NOTE: In the current implementation it assumes a constant Reynolds number on each airfoil throughout operation as captured in the polar used for initializing the FLOWVLM Rotor. The implementation of a varying Reynolds will be left for future development as needed.

FLOWVLM.OCCB2CCB — Method

Convert a rotor object from the old-CCBlade format to the new CCBlade type.

FLOWVLM.Vind — Method

Returns the velocity induced at point X

FLOWVLM._addsolution — Method

Extension of WingSystem's _addsolution() function

FLOWVLM._aveVinf — Method

Returns the average Vinf from all control points

FLOWVLM._calc_airfoils — Method

Calculates the airfoils at each control point

FLOWVLM._calc_distributedloads_lookuptable — Method

Calculates the load distribution by using the airfoil lookup table on the given inflow (this assumes that the inflow already includes all induced velocity and it is the effective inflow).

FLOWVLM._calc_inflow — Method

Returns the inflow velocity at the requested target point on all horseshoes, where the inflow is calculated as freestream + rotational velocity.

FLOWVLM._calculate_areas — Method

Calculates the area of each panel

FLOWVLM._calculate_forces — Method

Aerodynamic force calculated through Kutta-Joukowski theorem. Give it per_unit_span=true to calculate the force per unit length of span. If lifting_interac==false, the force doesn't include the induced velocity, hence, no interaction between lifting surfaces is calculated and induced drag will be zero, but it is an efficient way of calculating the lift of an individual lifting surface

FLOWVLM._calculate_moments — Method

Calculates total moment and decomposition into roll, pitch, and yaw

FLOWVLM._ccblade2global — Method

Receives a vector in CCBlade's coordinate system relative to blade and transforms it into the global coordinate system. NOTE: This function only rotates V into the new axis without translating it unless otherwise indicated.

FLOWVLM._ccbladeOaxis — Method

Returns the CCBlade's transformation matrix relative to the blade's c.s.

FLOWVLM._check — Method

Checks for consistency in internal variables

FLOWVLM._counter_interpret — Method

For a coordinate system 'inception2' that is incapsulated inside another coordinate system 'inception1', it receives its interpretation in the global coordinate system and counterinterprets it back to the 'inception1' system.

FLOWVLM._decompose — Method

Decomposes a force field into sideslip, drag, and lift components

FLOWVLM._fetch_wing — Method

Extension of WingSystem's _fetch_wing() function

FLOWVLM._fetch_wing — Method

Given a panel index in the system, it returns the wing the panel belongs to and that wing's index of the panel

FLOWVLM._generate_blade — Method

Generates the blade and discretizes it into lattices

FLOWVLM._get_O — Method

Extension of WingSystem's _get_O() function

FLOWVLM._get_Oaxis — Method

Extension of WingSystem's _get_O() function

FLOWVLM._global2ccblade — Method

Receives a vector in the global coordinate system and transforms it into CCBlade's coordinate system relative to blade. NOTE: This function only rotates V into the new axis without translating it unless otherwise indicated. (for definition of axes see notebook entry 20171202)

FLOWVLM._interpret — Method

For a coordinate system 'inception2' that is incapsulated inside another coordinate system 'inception1', it returns its interpretation in the global coordinate system

FLOWVLM._span_eff — Method

Returns the span effective of a Wing or WingSystem

FLOWVLM._verif_discr — Method

Verifies correct splining for lattice and element discretization

FLOWVLM.addchord — Method

addchord(wing, x, y, z, c, twist, n, r=1.0)

Adds a new chord to the wing and creates n lattices in the new section. Wing must be build from left to right.

Arguments

x       : x-position leading edge of the chord.

y       : y-position of the chord.

z       : z-position leading edge of the chord.

c       : Chord length.

twist   : Twist of the chord in degrees.

n::Int64: Number of lattices in the new section.

Optional arguments

r       : Ratio between lengths of first and last lattices.

central : Give it true to take the length ratio between the lattice
          midway and first and last. Give it a number between 0 and 1
          to define the position of the reference midway.

refinement : Use this option for more complex refinements. It
          receives an array `[sec1, sec2, ...]` with sections of
          refinements in the format `sec=[c, n, r]`, with `c` the length
          of this section (sum of all c = 1), `n` the ratio of lattices
          in this section, and `r` the increment ratio. If this option
          is used, it will ignore arguments `r` and `central`.

Examples

`julia> wing = Wing(0.0, 0.0, 0.0, 10.0, 3.0);`
`julia> addchord(wing, 2.5, 10.0, 5.0, 5.0, 0.0, 10);`

FLOWVLM.addwing — Method

Extension of WingSystem's addwing() function

FLOWVLM.addwing — Method

addwing(self::WingSystem, wing_name::String, wing::Wing) Adds a wing to the system with the position and orientation of local reference being interpreted in relation to the local reference frame of the system.

FLOWVLM.axis_rotation — Method

returns the transformation matrix of rotation around an arbitrary axis of unit vector r

FLOWVLM.calc_aerodynamicforces — Method

Calculates sectional aerodynamic forces in a rotor where the field DistributedLoad has already been solved for. It also calculates the bound circulation Gamma

FLOWVLM.calc_distributedloads — Method

Calculates the distributed loads from CCBlade. It also stores normal (Np) and tangential (Tp) components relative to the plane of rotation as given by CCBlade, if include_comps==true.

If return_performance==true, it returns propulsive efficiency eta, thrust coefficient CT, and torque coefficient CQ of each blade.

NOTE: These loads are per unit length of span

FLOWVLM.calc_inflow — Method

Receives the freestream velocity function V(x,t) and the current RPM of the rotor, and it calculates the inflow velocity field that each control point sees in the global coordinate system

FLOWVLM.calc_thrust_torque — Method

calc_thrust_torque(rotor)

Integrates the load distribution along every blade to return the thrust and torque of the rotor.

FLOWVLM.calc_thrust_torque_coeffs — Method

calc_thrust_torque_coeffs(rotor, rho, t)

Integrates the load distribution along every blade to return the thrust and torque coefficients of the rotor. Thrust coefficient CT calculated as CT= rac{T}{ ho n^2 D^4}, torque coefficient CQ calculated as CQ= rac{Q}{ ho n^2 D^5}

FLOWVLM.calculate_field — Method

calculate_field(wing, field_name [, qinf, S, l, r_cg]) Calculates the field field_name on a Wing or WingSystem. See FIELDS for a list of implemented fields.

FLOWVLM.calculate_force_trefftz — Method

Calculates the force induced by the wake sheet assuming it planar in the direction of Vinf. This method works well only for an isolated lifting surface and should not be used on interacting lifting surfaces.

FLOWVLM.check_coord_sys — Method

Checks that the unit vectors given as the matrix M=[i;j;k] define a coordinate system

FLOWVLM.complexWing — Method

`complexWing(b::Float64, AR::Float64, tr::Float64, n::Int64,
                    pos::Array{Float64,1}, twist::Array{Float64,1},
                    sweep::Array{Float64,1}, dihed::Array{Float64,1};
                    symmetric::Bool=true)`

Returns a Wing object made of multiple sections with the specified dimensions.

Arguments

b::Float64 : Span.

AR::Float64 : Aspect ratio (span over tip chord).

n::Int64 : Number of lattices in semi-span.

pos::Array{Float64,1} : Position of each chord along semi-span.

clen::Array{Float64,1} : Length of each chord as a fraction of tip chord.

twist::Array{Float64,1} : (deg) twist of each chord.

sweep::Array{Float64,1} : (deg) sweep of each section.

dihed::Array{Float64,1} : (deg) dihedral of each section.

Optional Arguments

symmetric::Bool=true : If false, returns only the semi-span.

chordalign::Float64=0.0 : Indicate position along chord length to align chords. Give it 0 for alignment about leading edge, 0.25 for alignment about quarter chord, and 1.0 for alignment about trailing edge.

Example

  # Wing Parameters
  b = 1.0                     # (m) span
  AR = 12.0                   # Span over tip chord
  n = 50                      # Lattices in semi-span

  # Chords
  pos = Float64[0, 0.25, 0.75, 1]  # Position of each chord along semi-span
  clen = Float64[2, 1.5, 1.5, 1]   # Length of each chord
  twist = Float64[0, 0, -2, -4]    # (deg) twist at each chord
  sweep = Float64[10, 15, 35]      # (deg) sweep of each section
  dihed = Float64[2, 5, 7.5]       # (deg) dihedral of each section

  # Generates the wing
  wing = vlm.complexWing(b, AR, n, pos, clen, twist, sweep, dihed)

FLOWVLM.copy — Method

Returns a deep copy of this wing

FLOWVLM.countertransform — Method

Rotates and translates back a vector V' that had been rotated and translated into the system (i', j', k') with origin T, and returns the original V. To ease repetitive computation, instead of giving the unit vectors, give the inverse of their matrix.

FLOWVLM.ellipticWing — Method

Returns a wing of elliptic load distribution (no sweep, no dihedral, no tapper). Give it elliptic="twist" to get an elliptic twist distribution, otherwise it will build an elliptic chord distribution.

FLOWVLM.exampleWing — Method

Example of a system of lifting surfaces (WingSystem)

FLOWVLM.fields_summary — Method

Returns a dictionary summarizing key parameters on a Wing or WingSystem

FLOWVLM.getControlPoint — Method

Returns the m-th control point of the system

FLOWVLM.getControlPoint — Method

Returns the m-th control point

FLOWVLM.getControlPoint — Method

Returns the m-th control point of the system

FLOWVLM.getHorseshoe — Method

Returns the m-th horseshoe of the system in the global coordinate system

FLOWVLM.getHorseshoe — Method

Returns the m-th horseshoe in the global coordinate system

FLOWVLM.getHorseshoe — Method

Returns the m-th horseshoe of the system in the global coordinate system

FLOWVLM.getHorseshoes — Method

Returns all the horseshoes of the Wing or WingSystem

FLOWVLM.getLE — Method

Returns leading-edge coordinates of the n-th chord

FLOWVLM.getTE — Method

Returns trailing-edge coordinates of the n-th chord

FLOWVLM.getVinfs — Method

Returns the undisturbed freestream at each control point

FLOWVLM.getVinfs — Method

Returns the undisturbed freestream at each control point, or at the horseshoe point indicated as target.

FLOWVLM.getVinfs — Method

Returns the undisturbed freestream at each control point, or at the horseshoe point indicated as target.

FLOWVLM.get_A — Method

Returns the area of the m-th panel, with self either Wing or WingSystem. For derivation, see notebook entry 20170619.

FLOWVLM.get_CG — Method

Returns the center of gravity of the wing, with self either Wing or WingSystem.

FLOWVLM.get_barc — Method

Returns the mean chord length of a Wing or WingSystem

FLOWVLM.get_blade — Method

Returns the requested blade

FLOWVLM.get_m — Method

Returns total number of lattices in the rotor

FLOWVLM.get_m — Method

Returns total number of lattices in the wing

FLOWVLM.get_m — Method

Returns total number of lattices in the wing

FLOWVLM.get_mBlade — Method

Returns total number of lattices on each blade

FLOWVLM.get_r — Method

Returns the center of gravity of the m-th panel, with self either Wing or WingSystem.

FLOWVLM.get_wing — Method

Extension of WingSystem's get_wing() function

FLOWVLM.get_wing — Method

Returns the i-th wing in the system

FLOWVLM.get_wing — Method

Returns the wing in the system under the requested name

FLOWVLM.get_wings — Method

Returns the wings in the system as a dictionary Dict(wing_name => Wing)

FLOWVLM.grid_dependance — Method

Receives a function generate_wing(n) and plots aerodynamic characteristics at different lattice resolutions

FLOWVLM.initialize — Method

Initializes the geometry of the rotor, discretizing each blade into n lattices

FLOWVLM.occb_af_from_data — Method

af_from_data(alpha, cl, cd)

Create an AirfoilData object directly from alpha, cl, and cd arrays.

affromaerodynfile calls this function indirectly. Uses a cubic B-spline (if the order of the data permits it). A small amount of smoothing of lift and drag coefficients is also applied to aid performance for gradient-based optimization.

FLOWVLM.occb_airfoil — Method

private

evalute airfoil spline at alpha

FLOWVLM.planform_area — Method

Receives a Wing or WingsSystem and return the planform area, which is the area of the projection on the xy-plane of its local coordinate system.

FLOWVLM.rotate — Method

Rotates the rotor degs degrees in the direction of rotation

FLOWVLM.save — Method

Saves the rotor in VTK legacy format

FLOWVLM.save — Method

Saves the wing domain in VTK legacy format

FLOWVLM.save_loft — Method

Saves the lofted surface of the blade

FLOWVLM.setRPM — Method

Sets RPM as the revolutions per minutes of this rotor

FLOWVLM.setVinf — Method

Sets Vinf(X,t) as the incoming freestream of this rotor

FLOWVLM.setVinf — Method

Sets Vinf(X,t) as the freestream of this wing

FLOWVLM.setVinf — Method

Sets Vinf(X,t) as the freestream of all wings in the system

FLOWVLM.setcoordsystem — Method

Sets a coordinate system for the rotor. If the user is calling this function, give user=true, otherwise it won't do the automatic translation to blade c.s.

FLOWVLM.setcoordsystem — Method

setcoordsystem(self, O, Oaxis; check=true)

Redefines the local coordinate system of the wing, where O is the new origin and Oaxis is the matrix [i; j; k] of unit vectors

FLOWVLM.setcoordsystem — Method

setcoordsystem(self, O, Oaxis; check=true, wings=String[]) Redefines the local coordinate system of the system, where O is the new origin and Oaxis is the matrix [i; j; k] of unit vectors. It transforms the coordinates of all the wings in the system accordingly.

To change the local coordinate system of a specific wing relative to the system's coordinate system, give the name of the wing in an array under argument wings.

FLOWVLM.simpleWing — Method

`simpleWing(b, ar, tr, twist, lambda, gamma; twist_tip=twist, n=20, r=2.0)`

Generates a single-section wing.

Arguments

b : (float) span.

ar : (float) aspect ratio defined as the span over the tip's chord.

tr : (float) taper ratio (tip chord / root chord).

twist : (float) twist of the root in degrees.

lambda : (float) sweep in degrees.

gamma : (float) dihedral in degrees.

(OPTIONALS)

twist_tip : (float) twist of the tip if different than root.

n : (int) number of horseshoes per side of the wing.

r : (float) horseshoes' expansion ratio

FLOWVLM.solve — Method

Solves the VLM of the Wing or WingSystem

FLOWVLM.solvefromCCBlade — Method

Solves for the Gamma field (circulation) on each blade using CCBlade. It also includes the fields Ftot, L, D, and S. (WARNING: These Ftot, L, D, and S are forces per unit length!)

If include_comps==true it stores CCBlade-calculated normal and tangential forces in the Rotor.

FLOWVLM.solvefromV — Method

Given the velocity induced at each control point (Vind = Vliftsurface+Vwake), solves for the Gamma field (circulation) on each blade by looking at the airfoil polar at the effective angle of attack of every section. It also includes the fields Ftot, L, D, and S. (WARNING: These Ftot, L, D, and S are forces per unit length!)

THIS METHOD IS UNSTABLE.

NOTE: Vind is expected to be in the global coordinate system. NOTE: Vind is expected to be formated as Vind[i][j] being the velocity vector of the j-th control point in the i-th blade.

FLOWVLM.solvefromVite — Method

Given the velocity induced at each control point (Vind = Vwake, no lifting surface), solves for the Gamma field (circulation) on each blade by looking at the airfoil polar at the effective angle of attack of every section. It also includes the fields Ftot, L, D, and S. (WARNING: These Ftot, L, D, and S are forces per unit length!)

This method solves iteratively until the circulation distribution converges.

NOTE: Vind is expected to be in the global coordinate system. NOTE: Vind is expected to be formated as Vind[i][j] being the velocity vector of the j-th control point in the i-th blade.

FLOWVLM.transform — Method

Rotates and translates the vector V.

Receives the i', j', k' unit vectors of an euclidean system with origin T, and returns V'. (In this version, the unit vectors have been organized as a matrix M)

FLOWVLM.OCCBRotor — Type

Rotor(r, chord, theta, af, Rhub, Rtip, B, precone)

Define rotor geometry.

Arguments

r::Array{Float64, 1}: radial locations (m)

chord::Array{Float64, 1}: chord lengths (m)

theta::Array{Float64, 1}: total twist including pitch (rad)

af::Array{AirfoilData, 1}: airfoils

Rhub::Float64: hub radius (along blade length)

Rtip::Float64: tip radius (along blade length)

B::Int64: number of blades

precone::Float64: precone angle (rad)

FLOWVLM.Rotor — Type

Rotor(CW, r, chord, theta, LE_x, LE_z, B, airfoil)

Object defining the geometry of a rotor/propeller/wind turbine. This class behaves as an extension of the WingSystem class, hence all functions of WingSystem can be applied to a Rotor object.

Arguments

CW::Bool : True for clockwise rotation, false for CCW.

r::Array{Float64,1} : Radius position for the following variables.

chord::Array{Float64,1} : Chord length.

theta::Array{Float64,1} : Angle of attack (deg) from the rotor's plane of rotation.

LE_x::Array{Float64,1} : x-position of leading edge (positive is ahead of radial axis relative to rotation).

LE_z::Array{Float64,1} : z-position of leading edge (height from plane of rotation).

B::Int64 : Number of blades.

Optional Arguments

airfoils::Array{Tuple{Float64, airfoilprep.Polar},1} : 2D airfoil properties along blade in the form [ (ri, Polari) ] with Polari describes the airfoil at i-th radial position ri (both the airfoil geometry in Polari and ri must be normalized). At least root (r=0) and tip (r=1) must be given so all positions in between can be extrapolated. This properties are only used when calling CCBlade and for generating good loking visuals; ignore if only solving the VLM.

NOTE: r here is the radial position after precone is included in the geometry, hence the need of explicitely declaring LE_z.

PROPERTIES

sol : Contains solution fields specific for Rotor types. They are formated as sol[fieldname] = Dict( "fieldname" => outputfieldname, "fieldtype" => "scalar" or "vector", "fielddata" => data ) where data is an array data[i] = [val1, val2, ...] containing this field values (scalar or vector) of all control points in the i-th blade.

<!– NOTE TO SELF: r is the y-direction on a wing, hence, remember to build the blade from root in the direction of positive y. –>

FLOWVLM.Wing — Type

Wing(leftxl, lefty, leftz, leftchord, leftchordtwist)

Wing constructor that automatically discretizes the surface into lattices. The wing must we built from left to right. Chords are parallel to the zx-plane, leading edge in the direction of the -xaxis and trailing in the direction of the +xaxis (hence, the span is aligned with the y-axis).

<!– The i-th control point is located at (xm[i], _ym[i]), with left bound –> <!– vortex (xn[i], yn[i]) and right bound vortex (xn[i+1], _yn[i+1]) –>

Arguments

leftxl : x-position of leading edge of the left tip.

lefty : y-position of left tip.

`leftzl : z-position of leading edge of the left tip.

leftchord : Chord of the left tip.

leftchordtwist : Twist left tip's chord in degrees.

Example

julia julia> wing = Wing(0.0, 0.0, 0.0, 10.0, 3.0);

FLOWVLM.WingSystem — Type

WingSystem() Initiates a system of wings. All method applicable to a Wing object are applicable to a WingSystem. When solved, it will calculate the interaction between wings.