Functions Contained in MyVPM

MyVPM.ExaFMM_autotuning — Method

Automatically tunes the paremeters of the FMM algorithm for optimum performance at the currrent number of particles in the field. WARNING: It has been observed that some of these parameters (specially theta) bring instability into the VPM method.

MyVPM.FDconv_ExaFMM — Method

Calls ExaFMM to calculate induced velocity at every particle and corresponding convective term (convective term calculated through finite difference)

MyVPM.K2 — Method

Second regularization function found in Winckelmans and Leonard, 1993, Table 1

MyVPM.U_ExaFMM — Method

Calls ExaFMM to calculate induced velocity at every particle

MyVPM.Uphi — Method

Default velocity induced by the scalar-potential flow

MyVPM._Utot — Method

Total velocity at position X at the current time with the requested kernel.

MyVPM._fetch — Method

Returns the vector X containing the data requested in to_fetch from the n-th particle, where X[i] = self._p_field[to_fetch[i]].

Arguments

to_fetch contains the indexes of the data to fetch as defined in the data structure definition block at the begining of this module.

Examples

julia> _fetch(p_field, 10, [_xi, _y_i, _zi])
Returns current position of the 10th particle

MyVPM._nextstep_euler — Method

Steps the field forward in time by dt in a first-order Euler integration scheme.

MyVPM._nextstep_rk — Method

Steps the field forward in time by dt in a third-order low-storage Runge-Kutta integration scheme. See Notebook entry 20180105.

MyVPM._update — Method

Updates the current data of the n-th particle with X as self._p_field[n, index[i]] = X[i]

MyVPM._update_Gamma — Method

Updates the current vectorial vorticity of the n-th particle

MyVPM._update_x — Method

Updates the current position vector of the n-th particle

MyVPM.addparticle — Method

`addparticle(self, particle)`

Adds a particle to the field, where particle must follow the data structure defined at the begining of this module.

Examples

julia> addparticle(p_field, [x, y, z, gamma_x, gamma_y, gamma_z, sigma, vol])

MyVPM.callExaFMM — Method

Calls ExaFMM to solve for the vector potential field induced at every particle and the jacobian. The output has the following format

OUTPUT: npX(nparam+3+9*4) matrix OUT where OUT[i][1:nparam] is the input parameters of the i-th particle, OUT[i][nparam+1:nparam+3] is the vector potential at the position of the i-th particle, and OUT[i][nparam+3+1:nparam+3+9] is the linearized jacobian, OUT[i][nparam+3+9+1:nparam+3+9*2] is dJdx1, OUT[i][nparam+3+9*2+1:nparam+3+9*3] is dJdx2, OUT[i][nparam+3+9*3+1:nparam+3+9*4] is dJdx3, OUT[i][nparam+3+9*4+1:n_param+3+9*4+3] is the particle strength exchange.

NOTE: The particle order in OUT matches that of pfield.

MyVPM.conv — Method

`conv(p_field, n; previous)`

Calculates the convective term at each particle. Returns the Npx6 matrix M where M[i, 1:3] is the value of the convective term at the position of the i-th particle, M[i, 4:6] is the total velocity at the position of the i-th particle. If vorticity==true it will M will be Npx9 with M[i, 7:9] the vorticity at the position of the i-th particle.

MyVPM.conv_ExaFMM — Method

Calls ExaFMM to calculate induced velocity at every particle and corresponding convective term

MyVPM.curlU — Method

Vorticity at X

MyVPM.delparticle — Method

`delparticle(self, n)`

Deletes the n-th particle in the field.

MyVPM.error_divergencefree_Uw — Method

Returns the sum of vector-potential velocity gradients over each particle's position

MyVPM.error_divergencefree_w — Method

Returns the sum of vorticity gradients over each particle's position

MyVPM.get_gamma — Method

Returns the vectorial circulation of the n-th particle

MyVPM.get_np — Method

Returns the current number of particles

MyVPM.get_particle — Method

Returns the n-th particle

MyVPM.get_sigma — Method

Returns the blob radius of the n-th particle

MyVPM.get_vol — Method

Returns the blob radius of the n-th particle

MyVPM.get_x — Method

Returns the position vector of the n-th particle

MyVPM.lagrangian_distortion — Method

Adds particles to the field for fixing Lagrangian distortion and runs an RBF for finding the new strengths in the field.

MyVPM.nextstep — Method

`nextstep(p_field, dt)`

Steps the field forward in time by dt. Use Uprobes to probe the velocity at any set of points. Give it relax=true to relax the discretized vorticity field back to a divergence-free field in this step.

MyVPM.rbf_cg — Method

Radial basis function interpolation of Gamma using the conjugate gradient method. This method only works on a particle field with uniform smoothing radius sigma.

MyVPM.read — Method

Reads and returns a particle field .jld

MyVPM.read_vtk — Method

Reads and returns a particle field from a .vtk

MyVPM.reg_UomegaBarba — Method

Direct calculation of induced velocity as a regularized Biot-Savart kernel using a Gaussian regularization

MyVPM.reg_UomegaGharakhani — Method

Direct calculation of induced velocity as a regularized Biot-Savart kernel using the Gharakhani's Gaussian regularization

MyVPM.reg_UomegaWinckelmann — Method

Direct calculation of induced velocity as a regularized Biot-Savart kernel using Winckelmann's regularization

MyVPM.reg_Utot — Method

Total velocity at position X at the current time with the singular kernel

MyVPM.run_vpm! — Method

run_vpm!(p_field, dt, nsteps; runtime_function=nothing,
              save_path=nothing, run_name="pfield", nsteps_save=1,
              save_code="", prompt=true)

Solves nsteps of the particle field with a time step of dt.

Optional Arguments

runtime_function::Any : Give it a function of the form myfun(pfield, t, dt). On each time step it will call this function. Use this for adding particles, deleting particles, etc.

solver_method::String : Solver method (ExaFMM, direct, directblob).

nsteps_relax::Int64 : Relaxes the particle field every this many time steps.

save_path::Any : Give it a string for saving VTKs of the particle field. Creates the given path.

run_name::String : Suffix of vtk files.

nsteps_save::Int64 : Saves vtks every this many time steps.

save_code::String : Give it the name of a file and it will save make a copy of this file along with the vtks. Use this for backing up whatever code is used for generating this run.

prompt::Bool : If save_path already exist, it will prompt the user before overwritting the folder if true; it will directly overwrite it if false.

verbose::Bool : Prints progress of the run to the terminal.

verbose_nsteps::Bool: Number of time steps between verbose.

parallel_Uphi::Bool : Flag for evaluating Uphi in parallel.

group_Uphi::Bool : Flag for evaluating Uphi as a group of points.

beta::Real : Maximum core growth (sgm/sgm0) in CS scheme.

sgm0::Real : Reinitialization core size.

MyVPM.save — Method

Outputs the particle field as the file [file_name].jld

MyVPM.save_vtk — Method

save_field(p_field, filename; path="", comment="")

Saves the particle field p_field as a vtk file.

MyVPM.sing_Uomega — Method

Direct calculation of induced velocity as a singular Biot-Savart kernel

MyVPM.sing_Utot — Method

Total velocity at position X at the current time with the singular kernel

MyVPM.zeta_ExaFMM — Method

Calls ExaFMM to calculate discretization function zeta at every particle

MyVPM.ParticleField — Type

ParticleField(n_max::Int64, Uinf, Uphi, method::String)

Initializes a particle field where up to n_max particles can be added and computed, the undisturbed freestream is given by Uinf(X, t), and a scalar-potential induced velocity is given by Uphi(X, t). method defines the solver method ("ExaFMM"/"direct"/"directblob")