UJ Scheme

FLOWVPM.Kernel — Type

Kernel(zeta, g, dgdr, g_dgdr, EXAFMM_P2P, EXAFMM_L2P)

Arguments

zeta::Function : Basis function zeta(r).

g::Function : Regularizing function g(r).

dgdr::Function : Derivative of g(r).

g_dgdr::Function : Efficient evaluation of g and dgdr.

EXAFMM_P2P::Int : Flag for the ExaFMM P2P function to call.

EXAFMM_L2P::Int : Flag for the ExaFMM L2P function to call.

FLOWVPM.UJ_direct — Function

UJ_direct(pfield)

Calculates the velocity and Jacobian that the field exerts on itself by direct particle-to-particle interaction, saving U and J on the particles.

NOTE: This method accumulates the calculation on the properties U and J of every particle without previously emptying those properties.

UJ_direct(source, target)

Calculates the velocity and Jacobian that the field source exerts on every particle of field target, saving U and J on the particles.

NOTE: This method accumulates the calculation on the properties U and J of every particle without previously emptying those properties.

FLOWVPM.UJ_fmm — Function

UJ_fmm(pfield)

Calculates the velocity and Jacobian that the field exerts on itself through a fast-multipole approximation, saving U and J on the particles.

NOTE: This method accumulates the calculation on the properties U and J of every particle without previously emptying those properties.

Missing docstring.

Missing docstring for FLOWUnsteady.vpm.FMM. Check Documenter's build log for details.