Utilities

FLOWVPM.run_vpm! — Function

run_vpm!(pfield, dt, nsteps; runtime_function=nothing, save_path=nothing, run_name="pfield", nsteps_save=1, verbose=true, prompt=true)

Solves nsteps of the particle field with a time step of dt.

Optional Arguments

runtime_function::Function : Give it a function of the form myfun(pfield, t, dt). On each time step it will call this function. Use this for adding particles, deleting particles, etc.

static_particles_function::Function : Give it a function of the form myfun(pfield, t, dt) to add static particles representing solid boundaries to the solver. This function is called at every time step right before solving the governing equations, and any new particles added by this function are immediately removed.

save_path::String : Give it a string for saving VTKs of the particle field. Creates the given path.

run_name::String : Name of output files.

nsteps_save::Int64 : Saves vtks every this many time steps.

prompt::Bool : If save_path already exist, it will prompt the user before overwritting the folder if true; it will directly overwrite it if false.

verbose::Bool : Prints progress of the run to the terminal.

verbose_nsteps::Bool: Number of time steps between verbose.

FLOWVPM.save — Function

save(pfield, file_name; path="")

Saves the particle field in HDF5 format and a XDMF file especifying its the attributes. This format can be opened in Paraview for post-processing and visualization.

Missing docstring.

Missing docstring for FLOWUnsteady.vpm.read. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FLOWUnsteady.vpm.save_settings. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FLOWUnsteady.vpm.read_settings. Check Documenter's build log for details.