Viscous Scheme

FLOWVPM.ViscousScheme — Type

`ViscousScheme{R}`

Type declaring viscous scheme.

Implementations must have the following properties: * nu::R : Kinematic viscosity.

Missing docstring.

Missing docstring for FLOWUnsteady.vpm.Inviscid. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FLOWUnsteady.vpm.CoreSpreading. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FLOWUnsteady.vpm.ParticleStrengthExchange. Check Documenter's build log for details.

FLOWVPM.rbf_conjugategradient — Function

Radial basis function interpolation of Gamma using the conjugate gradient method. This method only works on a particle field with uniform smoothing radius sigma.

See 20180818 notebook and https://en.wikipedia.org/wiki/Conjugategradientmethod#Theresultingalgorithm